requirements.txt
aniso8601==7.0.0
asgiref==3.2.7
astroid==2.3.3
attrs==19.3.0
autopep8==1.5
Django==3.0.4
graphene==2.1.8
graphql-core==2.3.1
graphql-relay==2.0.1
importlib-metadata==1.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
promise==2.3
py==1.8.1
pycodestyle==2.5.0
pylint==2.4.4
pyparsing==2.4.6
pytest==5.4.1
pytz==2019.3
Rx==1.6.1
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
wcwidth==0.1.9
wrapt==1.11.2
zipp==3.1.0
